fix(qua_parser): use u32 for int

+++ qua_parser/src/ast/literal.rs
@@ -3,5 +3,5 @@ pub enum Literal {
     String(String),
     Char(char),
     Float(f32),
-    Int(f32),
+    Int(u32),
 }
